fix: lint and move mixins to /mixins
