+++ frontend/src/mixins/search-ctl.ts
@@ -3,19 +3,16 @@ import {property, state} from 'lit/decorators.js';
 import {
   EventRegistrationInterface,
   EventRegistrationMixin,
-} from './event-listener-setup';
-import {SearchaliciousEvents} from './enums';
+} from '../event-listener-setup';
+import {SearchaliciousEvents} from '../enums';
 import {
   ChangePageEvent,
   LaunchSearchEvent,
   SearchResultEvent,
   SearchResultDetail,
-} from './events';
-import {SearchaliciousFacets} from './search-facets';
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-type Constructor<T = {}> = new (...args: any[]) => T;
-
+} from '../events';
+import {SearchaliciousFacets} from '../search-facets';
+import {Constructor} from './utils';
 export interface SearchaliciousSearchInterface
   extends EventRegistrationInterface {
   query: string;
