+++ frontend/src/mixins/utils.ts
@@ -1,1 +1,2 @@
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
 export type Constructor<T = {}> = new (...args: any[]) => T;
